Test: Replace test separators with something that stands out more
This helps when we're running the tests in a more verbose mode,
where pytest prints big bold "=====" strings.

We could change this with the `pytest_sessionstart(session)` hook
that looks at a command line argument we add to pytest and prints a
message based on that, but that seems like more work than it's worth.